% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CongestionDetection.R
\name{Data Driven congestion detection.}
\alias{Data Driven congestion detection.}
\title{Data driven congestion detection.}
\arguments{
\item{dataset}{including speed signal, change point probabilities}
}
\value{
Spatiotemporal congestions

classified dataframe
}
\description{
Data driven congestion detection.
}
\examples{
\dontrun{
library("Dcongestion")
dataset <- load(package = "Dcongestion", "dataset")
Congestion_detection(dataset)}
}
\author{
Atousa zarindast
}
