% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delay}
\alias{delay}
\title{Delay calculation.}
\usage{
delay(x, len)
}
\arguments{
\item{dataset}{including travel time, reference speed and actual speed}
}
\value{
Delay value

Calculated delay value
}
\description{
Delay calculation.
}
\examples{
\dontrun{
library("dcongestion")
visual <- load(package = "Dcongestion", "visual")
delay(visual)}
}
\author{
Atousa Zarindast
}
