% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_finder.R
\name{Elbow finder}
\alias{Elbow finder}
\alias{elbow_finder}
\title{Elbow finder.}
\usage{
elbow_finder(x_values, y_values)
}
\arguments{
\item{two}{dimentional space}
}
\value{
Elbow cut of point

cut of point
}
\description{
Elbow finder.
}
\author{
Atousa Zarindast
}
