% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_finder.R
\name{elbow_finder}
\alias{elbow_finder}
\title{Finds the elbow cut point.}
\usage{
elbow_finder(x_values, y_values)
}
\value{
Elbow cut of point

cut of point
}
\description{
Finds the elbow cut point.
}
\author{
Atousa Zarindast
}
