% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elbow_visual.R
\name{elbow_viusal}
\alias{elbow_viusal}
\title{Visualizing the elbow cut off point}
\usage{
elbow_viusal(visual)
}
\value{
congestion probabalities
}
\description{
Visualizing the elbow cut off point
}
\author{
Atousa zarindast
}
