% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delay}
\alias{delay}
\title{Delay calculation.}
\usage{
delay(x, len)
}
\arguments{
\item{dataset}{including travel time, reference speed and actual speed}
}
\value{
Delay value

A balanced dataframe
}
\description{
Delay calculation.
}
\examples{
\dontrun{
library("SentiAnalyzer")
direction <- system.file(package = "SentiAnalyzer", "extdata/Imbalance_Restaurant_Reviews.tsv")
imbalance_data<- read.delim(direction,quote='',stringsAsFactors = FALSE)
BalanceData(imbalance_data)}
}
\author{
Atousa Zarindast
}
