% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elbow_visual.R
\name{elbow_viusal}
\alias{elbow_viusal}
\title{Visualizing the elbow cut off point}
\usage{
elbow_viusal(visual)
}
\value{
balanced_dataframe balanced dataframe containing two columns: review texts and binary class , label: negative =0 and positive=1.

A balanced dataframe
}
\description{
Visualizing the elbow cut off point
}
\author{
Atousa zarindast
}
